:art: improve structure
